src/xtts_stream/api/broker/requirements.txt
fastapi
uvicorn
websockets
httpx
pyyaml
